finish the build part
