+++ packages/compass/package.json
@@ -161,6 +161,7 @@
     "download": "hadron-build download",
     "evergreen-expansions": "hadron-build info --format=yaml --flatten > expansions.yml",
     "build-info": "hadron-build info",
+    "write-assets": "hadron-build write-assets",
     "jsx": "babel ./src --out-dir ./lib",
     "eslint": "eslint",
     "prettier": "prettier",
