+++ packages/hadron-build/cli.js
@@ -15,6 +15,7 @@ const yargs = require('yargs')
   .command(require('./commands/download'))
   .command(require('./commands/ui'))
   .command(require('./commands/verify'))
+  .command(require('./commands/write-assets'))
   .demand(1, 'Please specify a command.')
   .strict()
   .env()
