+++ packages/hadron-build/test/fixtures/hadron-app/package.json
@@ -10,17 +10,27 @@
     "hadron": {
       "build": {
         "win32": {
-          "icon": "resources/win32/hadron-app.ico",
+          "icon": {
+            "dev": "resources/win32/hadron-app.ico"
+          },
           "favicon_url": "https://raw.githubusercontent.com/mongodb-js/favicon/master/favicon.ico",
-          "loading_gif": "src/app/images/win32/hadron-app-installer-loading.gif"
+          "loading_gif": "src/app/images/win32/hadron-app-installer-loading.gif",
+          "background": "src/app/images/win32/background.png",
+          "banner": "src/app/images/win32/banner.png"
         },
         "darwin": {
-          "icon": "resources/darwin/hadron-app.icns",
+          "icon": {
+            "dev": "resources/darwin/hadron-app.icns"
+          },
           "dmg_background": "resources/darwin/background.png",
           "app_category_type": "public.app-category.productivity"
         },
         "linux": {
-          "icon": "resources/linux/Icon.png"
+          "icon": {
+            "dev": "resources/linux/Icon.png"
+          },
+          "deb_section": "productivity",
+          "rpm_categories": ["IDE"]
         }
       },
       "distributions": {
